requirements.txt
Django==1.6
django-geoposition==0.1.5
django-simplesearch==1.0
MySQL-python==1.2.5
Pillow==3.0.0
requests==2.0.1
South==0.8.4
wheel==0.24.0
sorl-thumbnail==12.2
